[refactor, feat]: btn container folder renamed, showless event added
